flask-kafka-es-demo/requirements.txt
blinker==1.9.0
certifi==2025.10.5
click==8.3.0
elastic-transport==8.17.1
elasticsearch==8.15.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python==2.2.15
MarkupSafe==3.0.3
PyYAML==6.0.3
urllib3==2.5.0
Werkzeug==3.1.3
